# Hackathon2-phase3/specs/001-ai-todo-chatbot/contracts/chat-openapi.yaml
openapi: 3.0.3
info:
  title: AI Todo Chatbot API
  description: API for the AI-powered conversational todo management system
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: https://api.example.com/api
    description: Production server
paths:
  /{user_id}/chat:
    post:
      summary: Send a message to the AI chatbot and receive a response
      description: Stateless endpoint that processes user input and returns AI response with updated conversation history
      operationId: chat_with_agent
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the authenticated user
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Successful response from the AI agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request parameters or content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or expired JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user_id in path does not match authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ChatRequest:
      type: object
      required:
        - message
        - conversation_id
      properties:
        message:
          type: string
          description: The user's message to the AI agent
          example: "Add a task to buy groceries"
        conversation_id:
          type: string
          format: uuid
          description: The ID of the conversation to continue, or null for new conversation
          example: "123e4567-e89b-12d3-a456-426614174000"
    ChatResponse:
      type: object
      required:
        - response
        - conversation_id
        - message_id
      properties:
        response:
          type: string
          description: The AI agent's response to the user
          example: "I've added the task 'buy groceries' to your list."
        conversation_id:
          type: string
          format: uuid
          description: The ID of the conversation
          example: "123e4567-e89b-12d3-a456-426614174000"
        message_id:
          type: string
          format: uuid
          description: The ID of the AI's response message
          example: "987e6543-e21b-32d3-a456-426614174999"
        tool_calls:
          type: array
          items:
            $ref: '#/components/schemas/ToolCall'
          description: List of tools called during the interaction
    ToolCall:
      type: object
      required:
        - name
        - parameters
        - result
      properties:
        name:
          type: string
          description: Name of the tool that was called
          example: "add_todo"
        parameters:
          type: object
          description: Parameters passed to the tool
          example:
            title: "Buy groceries"
            description: "Get milk, bread, and eggs"
        result:
          type: object
          description: Result returned by the tool
          example:
            success: true
            todo_id: "abc123def456"
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "AUTH_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid authentication token"